주석 처리.

+++ Chap04/ProAss06.c
@@ -14,18 +14,23 @@
 #define _CRT_SECURE_NO_WARNINGS
 #include <stdio.h>
 
-
+ /*
+ * 함수명 : solve()
+ * 기능(책임) : 두 점의 좌표, 즉 (x1, y1)과 (x2, y2)
+ *				값을 입력받아 기울기를 계산한다.
+ * 반환 : 없음
+ */
 void solve()
 {
-	int x1, y1, x2, y2 = 0;
+	float x1, y1, x2, y2 = 0;
 
 	printf("한 점의 좌표 (x1, y1)?");
-	scanf("%d %d", &x1, &y1);
+	scanf("%f %f", &x1, &y1);
 
 	printf("또 다른 점의 좌표 (x2, y2)?");
-	scanf("%d %d", &x2, &y2);
+	scanf("%f %f", &x2, &y2);
 
-	printf("기울기 : %f\n", (float)(y2 - y1) / (x2 - x1));
+	printf("기울기 : %f\n", (y2 - y1) / (x2 - x1));
 }
 
 int main()
